+++ src/main/java/org/insset/client/exemple/ExemplePresenter.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.insset.client.addition;
+package org.insset.client.exemple;
 
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.event.dom.client.ClickEvent;
